+++ NEWS.md
@@ -9,6 +9,8 @@ cvc5 1.0.9
            disambiguation. The new syntax for 0-ary tuple sort is `UnitTuple`
            whose 0-ary constructor is `tuple.unit` (the previous syntax had
            overloaded `Tuple` and `tuple`, with no arguments).
+- API: Add the ability to query the logic that has been set in the solver via
+       `Solver::isLogicSet` and `Solver::getLogic`.
 
 cvc5 1.0.8
 ==========